requirements.txt
Flask
beautifulsoup4
requests
newspaper3k
lxml_html_clean
